+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/security/SalesforceKeyGenerator.java
@@ -27,37 +27,25 @@
 package com.salesforce.androidsdk.security;
 
 import android.content.SharedPreferences;
-import android.os.Build;
-import android.security.KeyPairGeneratorSpec;
-import android.security.keystore.KeyGenParameterSpec;
-import android.security.keystore.KeyProperties;
 import android.text.TextUtils;
 import android.util.Base64;
 
 import com.salesforce.androidsdk.app.SalesforceSDKManager;
 import com.salesforce.androidsdk.util.SalesforceSDKLogger;
 
-import java.io.IOException;
-import java.math.BigInteger;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
-import java.security.KeyPairGenerator;
-import java.security.KeyStore;
-import java.security.KeyStoreException;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.security.SecureRandom;
-import java.security.cert.CertificateException;
-import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 import java.util.UUID;
 
 import javax.crypto.KeyGenerator;
-import javax.security.auth.x500.X500Principal;
 
 /**
  * This class provides methods to generate a unique ID that can be used as an encryption
@@ -76,8 +64,6 @@ public class SalesforceKeyGenerator {
     private static final String SHA256 = "SHA-256";
     private static final String SHA1PRNG = "SHA1PRNG";
     private static final String AES = "AES";
-    private static final String RSA = "RSA";
-    private static final String ANDROID_KEYSTORE = "AndroidKeyStore";
 
     private static Map<String, String> UNIQUE_IDS = new HashMap<>();
     private static Map<String, String> CACHED_ENCRYPTION_KEYS = new HashMap<>();
@@ -151,57 +137,36 @@ public class SalesforceKeyGenerator {
     }
 
     /**
-     * Generates a keypair and returns the public key.
+     * Generates an RSA keypair and returns the public key.
      *
      * @param name Alias of the entry in which the generated key will appear in Android KeyStore.
+     * @param length Key length.
      * @return RSA public key.
      */
     public static synchronized PublicKey getRSAPublicKey(String name, int length) {
-        PublicKey publicKey = null;
-        createRSAKeysIfNecessary(name, length);
-        try {
-            publicKey = loadKeyStore().getCertificate(name).getPublicKey();
-        } catch (Exception e) {
-            SalesforceSDKLogger.e(TAG, "Security exception thrown", e);
-        }
-        return publicKey;
+        return KeyStoreWrapper.getInstance().getRSAPublicKey(name, length);
     }
 
     /**
-     * Generates a keypair and returns the encoded public key string.
+     * Generates an RSA keypair and returns the encoded public key string.
      *
      * @param name Alias of the entry in which the generated key will appear in Android KeyStore.
+     * @param length Key length.
      * @return RSA public key string.
      */
     public static synchronized String getRSAPublicString(String name, int length) {
-        final PublicKey publicKey = getRSAPublicKey(name, length);
-        String publicKeyBase64 = null;
-        if (publicKey != null) {
-            publicKeyBase64 = Base64.encodeToString(publicKey.getEncoded(),
-                    Base64.NO_WRAP | Base64.NO_PADDING);
-        }
-        return publicKeyBase64;
+        return KeyStoreWrapper.getInstance().getRSAPublicString(name, length);
     }
 
     /**
-     * Generates a keypair and returns the private key.
+     * Generates an RSA keypair and returns the private key.
      *
      * @param name Alias of the entry in which the generated key will appear in Android KeyStore.
+     * @param length Key length.
      * @return RSA private key.
      */
     public static synchronized PrivateKey getRSAPrivateKey(String name, int length) {
-        PrivateKey privateKey = null;
-        createRSAKeysIfNecessary(name, length);
-        try {
-            KeyStore.Entry entry = loadKeyStore().getEntry(name, null);
-            if (entry == null) {
-                return null;
-            }
-            privateKey = ((KeyStore.PrivateKeyEntry) entry).getPrivateKey();
-        } catch (Exception e) {
-            SalesforceSDKLogger.e(TAG, "Security exception thrown", e);
-        }
-        return privateKey;
+        return KeyStoreWrapper.getInstance().getRSAPrivateKey(name, length);
     }
 
     private static void generateEncryptionKey(String name) {
@@ -258,54 +223,5 @@ public class SalesforceKeyGenerator {
     private static String getAddendum(String name) {
         final String suffix = TextUtils.isEmpty(name) ? "" : name;
         return String.format(Locale.US, ADDENDUM, suffix);
-    }
-
-    private static void createRSAKeysIfNecessary(String name, int length) {
-        try {
-            KeyStore keyStore = loadKeyStore();
-            if (!keyStore.containsAlias(name)) {
-
-                // Generates a new key pair.
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                    KeyPairGenerator kpg = KeyPairGenerator.getInstance(
-                            KeyProperties.KEY_ALGORITHM_RSA, ANDROID_KEYSTORE);
-                    kpg.initialize(new KeyGenParameterSpec.Builder(
-                            name,
-                            KeyProperties.PURPOSE_ENCRYPT | KeyProperties.PURPOSE_DECRYPT)
-                            .setKeySize(length)
-                            .setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_RSA_PKCS1)
-                            .build());
-                    kpg.generateKeyPair();
-                } else {
-
-                    /*
-                     * TODO: Remove the 'else' block once minVersion > 23.
-                     */
-                    Calendar start = Calendar.getInstance();
-                    Calendar end = Calendar.getInstance();
-                    end.add(Calendar.YEAR, 30);
-                    KeyPairGeneratorSpec spec = new KeyPairGeneratorSpec.Builder(SalesforceSDKManager.getInstance().getAppContext())
-                            .setAlias(name)
-                            .setSubject(new X500Principal("CN=" + name))
-                            .setSerialNumber(BigInteger.TEN)
-                            .setStartDate(start.getTime())
-                            .setEndDate(end.getTime())
-                            .setKeySize(length)
-                            .build();
-                    KeyPairGenerator kpg = KeyPairGenerator.getInstance(RSA, ANDROID_KEYSTORE);
-                    kpg.initialize(spec);
-                    kpg.generateKeyPair();
-                }
-            }
-        } catch (Exception e) {
-            SalesforceSDKLogger.e(TAG, "Security exception thrown", e);
-        }
-    }
-
-    private static KeyStore loadKeyStore() throws CertificateException, NoSuchAlgorithmException,
-            IOException, KeyStoreException {
-        KeyStore keyStore = KeyStore.getInstance(ANDROID_KEYSTORE);
-        keyStore.load(null);
-        return keyStore;
     }
 }
