+++ libs/test/SalesforceSDKTest/src/com/salesforce/androidsdk/security/KeyStoreWrapperTest.java
@@ -26,6 +26,73 @@
  */
 package com.salesforce.androidsdk.security;
 
+import android.app.Application;
+import android.app.Instrumentation;
+
+import com.salesforce.androidsdk.TestForceApp;
+import com.salesforce.androidsdk.analytics.security.Encryptor;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import java.security.PrivateKey;
+import java.security.PublicKey;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.filters.SmallTest;
+import androidx.test.platform.app.InstrumentationRegistry;
+
+/**
+ * Tests for {@link KeyStoreWrapper}.
+ *
+ * @author bhariharan
+ */
+@RunWith(AndroidJUnit4.class)
+@SmallTest
 public class KeyStoreWrapperTest {
 
+    private static final String KEY_1 = "key_1";
+    private static final String KEY_2 = "key_2";
+
+    @Before
+    public void setUp() throws Exception {
+        final Application app = Instrumentation.newApplication(TestForceApp.class,
+                InstrumentationRegistry.getInstrumentation().getTargetContext());
+        InstrumentationRegistry.getInstrumentation().callApplicationOnCreate(app);
+    }
+
+    @Test
+    public void testGetRSAPublicString() {
+        final KeyStoreWrapper keyStoreWrapper = KeyStoreWrapper.getInstance();
+        Assert.assertNotNull("KeyStoreWrapper instance should not be null", keyStoreWrapper);
+        final String key1 = keyStoreWrapper.getRSAPublicString(KEY_1, 2048);
+        final String key1Again = keyStoreWrapper.getRSAPublicString(KEY_1, 2048);
+        final String key2 = keyStoreWrapper.getRSAPublicString(KEY_2, 2048);
+        Assert.assertEquals("Public keys with the same name should be the same", key1, key1Again);
+        Assert.assertNotSame("Public keys with different names should be different", key1, key2);
+    }
+
+    @Test
+    public void testGetRSAPrivateKey() {
+        final KeyStoreWrapper keyStoreWrapper = KeyStoreWrapper.getInstance();
+        Assert.assertNotNull("KeyStoreWrapper instance should not be null", keyStoreWrapper);
+        final PrivateKey key1 = keyStoreWrapper.getRSAPrivateKey(KEY_1, 2048);
+        final PrivateKey key1Again = keyStoreWrapper.getRSAPrivateKey(KEY_1, 2048);
+        Assert.assertEquals("Private keys with the same name should be the same", key1, key1Again);
+    }
+
+    @Test
+    public void testRSAEncryptDecrypt() {
+        final KeyStoreWrapper keyStoreWrapper = KeyStoreWrapper.getInstance();
+        Assert.assertNotNull("KeyStoreWrapper instance should not be null", keyStoreWrapper);
+        final PrivateKey privateKey = keyStoreWrapper.getRSAPrivateKey(KEY_1, 2048);
+        final PublicKey publicKey = keyStoreWrapper.getRSAPublicKey(KEY_1, 2048);
+        final String data = "Test data for encryption";
+        final String encryptedData = Encryptor.encryptWithRSA(publicKey, data);
+        Assert.assertNotSame("Encrypted data should not match original data", data, encryptedData);
+        final String decryptedData = Encryptor.decryptWithRSA(privateKey, encryptedData);
+        Assert.assertEquals("Decrypted data should match original data", data, decryptedData);
+    }
 }
