+++ libs/test/SalesforceSDKTest/src/com/salesforce/androidsdk/security/SalesforceKeyGeneratorTest.java
@@ -28,20 +28,17 @@ package com.salesforce.androidsdk.security;
 
 import android.app.Application;
 import android.app.Instrumentation;
-import androidx.test.platform.app.InstrumentationRegistry;
-import androidx.test.filters.SmallTest;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.salesforce.androidsdk.TestForceApp;
-import com.salesforce.androidsdk.analytics.security.Encryptor;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.security.PrivateKey;
-import java.security.PublicKey;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.filters.SmallTest;
+import androidx.test.platform.app.InstrumentationRegistry;
 
 /**
  * Tests for {@link SalesforceKeyGenerator}.
@@ -89,32 +86,5 @@ public class SalesforceKeyGeneratorTest {
         final String id2 = SalesforceKeyGenerator.getEncryptionKey(KEY_2);
         Assert.assertEquals("Encryption keys with the same name should be the same", id1, id1Again);
         Assert.assertNotSame("Encryption keys with different names should be different", id1, id2);
-    }
-
-    @Test
-    public void testGetRSAPublicString() {
-        final String key1 = SalesforceKeyGenerator.getRSAPublicString(KEY_1, 2048);
-        final String key1Again = SalesforceKeyGenerator.getRSAPublicString(KEY_1, 2048);
-        final String key2 = SalesforceKeyGenerator.getRSAPublicString(KEY_2, 2048);
-        Assert.assertEquals("Public keys with the same name should be the same", key1, key1Again);
-        Assert.assertNotSame("Public keys with different names should be different", key1, key2);
-    }
-
-    @Test
-    public void testGetRSAPrivateKey() {
-        final PrivateKey key1 = SalesforceKeyGenerator.getRSAPrivateKey(KEY_1, 2048);
-        final PrivateKey key1Again = SalesforceKeyGenerator.getRSAPrivateKey(KEY_1, 2048);
-        Assert.assertEquals("Private keys with the same name should be the same", key1, key1Again);
-    }
-
-    @Test
-    public void testRSAEncryptDecrypt() throws Exception {
-        final PrivateKey privateKey = SalesforceKeyGenerator.getRSAPrivateKey(KEY_1, 2048);
-        final PublicKey publicKey = SalesforceKeyGenerator.getRSAPublicKey(KEY_1, 2048);
-        final String data = "Test data for encryption";
-        final String encryptedData = Encryptor.encryptWithRSA(publicKey, data);
-        Assert.assertNotSame("Encrypted data should not match original data", data, encryptedData);
-        final String decryptedData = Encryptor.decryptWithRSA(privateKey, encryptedData);
-        Assert.assertEquals("Decrypted data should match original data", data, decryptedData);
     }
 }
